feat: se creo la clase alerta y utilizo en las vistas

+++ admin/actions/editar/editar_producto_acc.php
@@ -6,41 +6,47 @@ $file = $_FILES['imagen'] ?? false;
 
 try {
     $producto = (new producto())->filtradoId($data['id']);
-    $producto->eliminarSabores();
-    $producto->eliminarIngredientes();
+    if ($producto) {
+        $producto->eliminarSabores();
+        $producto->eliminarIngredientes();
 
-    if (isset($data['sabores'])) {
-        foreach ($data['sabores'] as $sabor_id) {
-            $producto->agregarSabores(intval($data['id']), intval($sabor_id));
+        if (isset($data['sabores'])) {
+            foreach ($data['sabores'] as $sabor_id) {
+                $producto->agregarSabores(intval($data['id']), intval($sabor_id));
+            }
         }
-    }
 
-    if (isset($data['ingredientes'])) {
-        foreach ($data['ingredientes'] as $ingrediente_id) {
-            $producto->agregarIngredientes(intval($data['id']), intval($ingrediente_id));
+        if (isset($data['ingredientes'])) {
+            foreach ($data['ingredientes'] as $ingrediente_id) {
+                $producto->agregarIngredientes(intval($data['id']), intval($ingrediente_id));
+            }
         }
-    }
 
-    if (!empty($file['tmp_name'])) {
-        $imagen = (new Imagen())->subirImagen(__DIR__ . "/../../../catalogo", $file);
-        (new Imagen())->borrarImagen(__DIR__ . "/../../../catalogo/" . $data['imagen_original']);
+        if (!empty($file['tmp_name'])) {
+            $imagen = (new Imagen())->subirImagen(__DIR__ . "/../../../catalogo", $file);
+            (new Imagen())->borrarImagen(__DIR__ . "/../../../catalogo/" . $data['imagen_original']);
+        } else {
+            $imagen = $data['imagen_original'];
+        }
+
+        $fechaActualizada = date('Y-m-d');
+
+        $producto->update(
+            Utilidades::sacarAcentos($data['titulo']),
+            $data['precio'],
+            Utilidades::sacarAcentos($data['descripcion']),
+            $imagen,
+            $data['categoria_id'],
+            $data['marca_id'],
+            $fechaActualizada,
+            $data['presentacion_id']
+        );
+        (new Alerta())->crearAlerta('success', "El producto <b>{$producto->getTitulo()}</b> se editó correctamente");
     } else {
-        $imagen = $data['imagen_original'];
+        (new Alerta())->crearAlerta('warning', "El producto que intenta editar no existe");
     }
-
-    $fechaActualizada = date('Y-m-d');
-
-    $producto->update(
-        Utilidades::sacarAcentos($data['titulo']),
-        $data['precio'],
-        Utilidades::sacarAcentos($data['descripcion']),
-        $imagen,
-        $data['categoria_id'],
-        $data['marca_id'],
-        $fechaActualizada,
-        $data['presentacion_id']
-    );
-    header('Location: ../../index.php?seccion=admin_productos');
+    header('Location: ../../?seccion=admin_productos');
 } catch (Exception $e) {
-    die("Ha ocurrido un error al editar el producto");
+    (new Alerta())->crearAlerta('danger', "El producto no se pudo editar");
+    header('Location: ../../?seccion=admin_productos');
 }
