+++ admin/actions/eliminar/eliminar_marca_acc.php
@@ -4,9 +4,15 @@ require_once "../../../functions/autoload.php";
 $id = $_GET['id'] ?? false;
 
 try {
-    $marca = (new Marca)->get_por_id($id);
-    $marca->delete();
-    header('Location: ../../index.php?seccion=admin_marcas');
+    $marca = (new Marca())->get_por_id($id);
+    if ($marca) {
+        $marca->delete();
+        (new Alerta())->crearAlerta('success', "La marca <b>{$marca->getNombre()}</b> se eliminó correctamente");
+    } else {
+        (new Alerta())->crearAlerta('warning', "La marca que intenta eliminar no existe");
+    }
+    header('Location: ../../?seccion=admin_marcas');
 } catch (Exception $e) {
-    die("Ha ocurrido un error al intentar eliminar la marca");
+    (new Alerta())->crearAlerta('danger', "La marca no se puede eliminar ya que se esta usando en los productos");
+    header('Location: ../../?seccion=admin_marcas');
 }
