+++ admin/actions/eliminar/eliminar_producto_acc.php
@@ -5,12 +5,17 @@ $id = $_GET['id'] ?? false;
 
 try {
     $producto = (new Producto())->filtradoId($id);
-    $producto->eliminarSabores();
-    $producto->eliminarIngredientes();
-    $producto->delete();
-    (new Imagen())->borrarImagen(__DIR__ . "/../../../catalogo/" . $producto->getImagen());
-    //(new Alerta())->add_alerta('danger', "El producto <strong>" . $producto->getTitulo() ."</strong> se eliminó correctamente");
-    header('Location: ../../index.php?seccion=admin_productos');
+    if ($producto) {
+        $producto->eliminarSabores();
+        $producto->eliminarIngredientes();
+        $producto->delete();
+        (new Imagen())->borrarImagen(__DIR__ . "/../../../catalogo/" . $producto->getImagen());
+        (new Alerta())->crearAlerta('success', "El producto <b>{$producto->getTitulo()}</b> se eliminó correctamente");
+    } else {
+        (new Alerta())->crearAlerta('warning', "El producto que intenta eliminar no existe");
+    }
+    header('Location: ../../?seccion=admin_productos');
 } catch (Exception $e) {
-    die("Ha ocurrido un error al intentar al eliminar el producto");
+    (new Alerta())->crearAlerta('danger', "Ha ocurrido un error al intentar al eliminar el producto");
+    header('Location: ../../?seccion=admin_productos');
 }
