+++ admin/actions/eliminar/eliminar_sabor_acc.php
@@ -4,9 +4,15 @@ require_once "../../../functions/autoload.php";
 $id = $_GET['id'] ?? false;
 
 try {
-    $sabor = (new Sabor)->get_por_id($id);
-    $sabor->delete();
-    header('Location: ../../index.php?seccion=admin_sabores');
+    $sabor = (new Sabor())->get_por_id($id);
+    if ($sabor) {
+        $sabor->delete();
+        (new Alerta())->crearAlerta('success', "El sabor <b>{$sabor->getNombre()}</b> se eliminó correctamente");
+    } else {
+        (new Alerta())->crearAlerta('warning', "El sabor que intenta eliminar no existe");
+    }
+    header('Location: ../../?seccion=admin_sabores');
 } catch (Exception $e) {
-    die("Ha ocurrido un error al intentar eliminar el sabor");
+    (new Alerta())->crearAlerta('danger', "El sabor no se puede eliminar ya que se esta usando en los productos");
+    header('Location: ../../?seccion=admin_sabores');
 }
